Teleportation obj and teleport code
I added teleportation object that is camoflaged (invisible), but you can change its material (color) to make it visible (if you do not encounter it).

With the teleportation object i added an empty object in the world to refer to both the teleporter and the player, making the teleportation work.